added topic detail page
