+++ app/src/main/AndroidManifest.xml
@@ -38,7 +38,8 @@
         <activity
             android:name=".EditAccountActivity"
             android:label="@string/title_activity_edit_account"
-            android:theme="@style/AppTheme.NoActionBar"></activity>
+            android:theme="@style/AppTheme.NoActionBar" />
+        <activity android:name=".TopicDetail"></activity>
     </application>
 
 </manifest>
