+++ app/src/main/java/com/example/smart/emotionanalyzer/MainActivity.java
@@ -27,6 +27,7 @@ public class MainActivity extends AppCompatActivity implements BottomNavigationV
         String fragment = getIntent().getExtras().getString("fragment");
         mAuth = FirebaseAuth.getInstance();
 
+
         BottomNavigationView navigation = (BottomNavigationView) findViewById(R.id.navigation);
         navigation.setOnNavigationItemSelectedListener(this);
         if(fragment.equals("browse")) {
@@ -56,8 +57,6 @@ public class MainActivity extends AppCompatActivity implements BottomNavigationV
         return true;
     }
 
-
-
     @Override
     public boolean onNavigationItemSelected(@NonNull MenuItem item) {
         Fragment fragment = null;
@@ -76,31 +75,6 @@ public class MainActivity extends AppCompatActivity implements BottomNavigationV
             fragment = new AccountFragment();
         }
         return loadMyFragment(fragment);
-    }
-
-    @Override
-    public boolean onCreateOptionsMenu(Menu menu) {
-        getMenuInflater().inflate(R.menu.menu_main, menu);
-        return true;
-    }
-
-    @Override
-    public boolean onOptionsItemSelected(MenuItem item) {
-        int id = item.getItemId();
-        switch (id){
-            case R.id.item1:
-                Bundle bundle = new Bundle();
-                bundle.putString("name", mAuth.getCurrentUser().getDisplayName());
-                Toast.makeText(getApplicationContext(), "Item selected", Toast.LENGTH_SHORT).show();
-                Intent intent = new Intent(MainActivity.this, CreateTopicActivity.class);
-                intent.putExtras(bundle);
-                startActivity(intent);
-
-                return true;
-
-            default:
-                return super.onOptionsItemSelected(item);
-        }
     }
 
 }
