+++ app/src/main/java/com/example/smart/emotionanalyzer/TopicDetail.java
@@ -1,0 +1,13 @@
+package com.example.smart.emotionanalyzer;
+
+import android.support.v7.app.AppCompatActivity;
+import android.os.Bundle;
+
+public class TopicDetail extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_topic_detail);
+    }
+}
