--- app/src/main/res/menu/menu_main.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    tools:context="MainActivity">
-
-    <item  android:id="@+id/item1"
-        android:title="Create Topic"/>
-</menu>
